+++ src/Ryujinx/Systems/Configuration/ConfigurationState.cs
@@ -80,6 +80,9 @@ namespace Ryujinx.Ava.Systems.Configuration
                 TickScalar = System.TickScalar,
                 EnableInternetAccess = System.EnableInternetAccess,
                 EnableSwitchNet = System.EnableSwitchNet,
+                SwitchNetServer = System.SwitchNetServer,
+                SwitchNetDeviceAccountId = System.SwitchNetDeviceAccountId,
+                SwitchNetPassword = System.SwitchNetPassword,
                 EnableFsIntegrityChecks = System.EnableFsIntegrityChecks,
                 FsGlobalAccessLogMode = System.FsGlobalAccessLogMode,
                 AudioBackend = System.AudioBackend,
@@ -218,6 +221,9 @@ namespace Ryujinx.Ava.Systems.Configuration
             System.EnablePtc.Value = true;
             System.EnableInternetAccess.Value = false;
             System.EnableSwitchNet.Value = false;
+            System.SwitchNetServer.Value = string.Empty;
+            System.SwitchNetDeviceAccountId.Value = string.Empty;
+            System.SwitchNetPassword.Value = string.Empty;
             System.EnableFsIntegrityChecks.Value = true;
             System.FsGlobalAccessLogMode.Value = 0;
             System.AudioBackend.Value = AudioBackend.SDL3;